requirements.txt
# Core dependencies
selenium==4.15.2
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0

# WebDriver management
webdriver-manager==4.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
selenium-wire==5.1.0

# Logging and utilities
structlog==23.2.0
colorama==0.4.6  # For colored console output on Windows

# CLI enhancements
click==8.1.7
rich==13.5.2

# Authentication and security
bcrypt==4.1.2

# Type checking (optional, for development)
mypy==1.7.1
types-requests==2.31.0.10

# Code quality (optional, for development)
black==23.12.0
flake8==6.1.0
isort==5.13.2